Changes made in relation to the heroku deployment
public file content, composer.lock and Procfile are all required by heroku and so had to be removed from the gitignore in order to be included in the heroku deployment. Heroku is set up to automatically be updated from this repos base branch. Sass and webpack files not NOT able to be compiled on the server, so need to be delivered precompiled