updates to login form, routes, auth controllers
changes force to login form to be displayed when the user is not signed in and attempts to access any site route. Controllers have now all got an auth constructor method, this forces auth to confirm the user is logged in. If they are not the app redirects to the login form.  Register link has been commented out from the main page to prevent anyone from signing up