Sprint 4 authentication system (#148)
* App starter files including authentication scaffolding added

Project files copied to this new app configuration. Views now need to be modified to return the app to its original presentation, but with the authentication system included.  Will have this done before the end of sprint 4

* Updates made to the view files structure to combine files from 2 projects

* Modified routes and controllers, so these are protected

Displaying login form as the index page

* updates to login form, routes, auth controllers

changes force to login form to be displayed when the user is not signed in and attempts to access any site route. Controllers have now all got an auth constructor method, this forces auth to confirm the user is logged in. If they are not the app redirects to the login form.  Register link has been commented out from the main page to prevent anyone from signing up

* Login function now completed and functional.

Still some styling issues and debugging needed to register a user. the register user form has not been fully integrated with our style sheet, but it does work with the backend database.

* Update to navbar font. Made this bolder so it is easier to read

* register page now updated and tested.

Also updated the nav so that the login link shows when on the register user page and register shows on the login page

* Removed copy of project from the directory, corrected nav bar issue with logout link

* Commented code removed as per code review feedback