Altered the user registration system, added error handling (#156)
User registration routes are now only accessible to currently logged in users. This will protect the back end by preventing registration by non-registered users. This is a temporary fix, will need to be revisited in the event this gets rolled out to students. Also, included a fallback route with will return a user to the homepage if they enter a url that does not exist. Since registration is heavily restricted, I will be adding a admin user to the seeders list so that there is always atleast one user that can log in to start off