Sprint 5 css update (#170)
* Update to the html in the footer and nav bar.

Changed the style sheet to make the nav bar logo larger. Added linked html in the footer so a user can click the email, phone numbers - will be useful on mobile devices if the app got deployed to students/users off campus

* Update to student page form html

updated sass, css to improve layout, stripped out excess classes and wrapper html elements that were not doing anything. Fixed the table structure and changed the edit button to a form element so that this action would have cross site @csrf protection

* Updates made to the cohort and student pages style/layout

Included changes made on the cohort page from a seperate branch so that I am styling the most update version of the layout structure

* Update to students page layout, Sass/css

Updates include altering the main tag that wraps in the inner content so that sections align in columns

* Update to students page layout, Sass/css

Updates include altering the main tag that wraps in the inner content so that sections align in columns

* Updates to all form html and css

all layouts have been affected by these changes which improve layout, utilise more screen real estate

* Updates to register and loging form css