Sprint 5 CSS layout updates (#189)
* Footer styling updated

flex blox used to allow the footer columns to change with screen size

* replaced css flexbox with css grid

still some issues to debug, the layout have improved with this change. Alot of code has been removed due to grid natively handling alot of the functions this code was performing

* updated student scroll area on students page

* Updates to css, removed some unused classes and ids

* updated layout css

reverted container back to flex, stripped out classes that were not being used. this seems to have helped fix some bugs

* Updates to grid layout - debugging issues

* added a 12 col grid - now need to arrange grid items

* Revert "updated layout css"

This reverts commit ff613c61a378a9484f5cf1d0054e5c7a906e82e4.

* Revert "Updates to css, removed some unused classes and ids"

* Updated layout html and css

needed to revert to previous commits to fix some issues

* updated html and css layout - ready to review

please merge this to base-beta-testing before merging this to base

* updated html and css layout - ready to review

please merge this to base-beta-testing before merging this to base

* updates to speed of carousel + styling updates

* fix footer icon hover states